{"schema":"decodex/commit/1","summary":"Modularize state runtime model records","authority":"XY-1115"}
